requirements.txt
numpy
locust
Pillow
qwen_vl_utils
torch
torchvision
tqdm